refactoring to make as equal as possible to playax repo
